refactor: add pool and init helpers to lib
